+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Namespace documentation for every parameter.
 - New feature: `beforeForFirstBuild`. For first build `before` parameters will be applied, for subsequent builds `watch` parameters will be applied.Works only in `watch` key. [#25](https://github.com/Amaimersion/remove-files-webpack-plugin/issues/25)
+- New feature: `skipFirstBuild`. First build will be skipped. Works only in `watch` key.
 
 ## Changed